requirements.txt
playwright
snownlp
pandas
matplotlib
tqdm
beautifulsoup4
streamlit
plotly
wordcloud
jieba
python-dotenv
xlsxwriter
pythainlp
nltk
